requirements.txt
streamlit>=1.28.0
langchain==0.3.25
langchain-community==0.3.24
langchain-google-genai==2.0.10
langchain-core==0.3.61
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-docx>=0.8.11
google-generativeai>=0.3.0
gspread>=5.0.0
google-auth>=2.0.0
pandas>=1.5.0